财报任务
git-svn-id: http://gurucaijing.googlecode.com/svn/trunk@870 910414aa-d3c9-269b-9558-6cae95c76e5e

+++ caijing/src/main/java/com/caijing/task/TidyFinancialReportTask.java
@@ -76,7 +76,6 @@ public class TidyFinancialReportTask {
 				File[] reportFiles = TidyFinancialReportTask.listFileBySuffix(reportDir, ".pdf");
 				if (reportFiles != null) {
 					for (File reportfile : reportFiles) {
-						
 						System.out.println("==> report: " + reportfile.getPath());
 						String report_title = reportfile.getName();
 						String stockcode = report_title.split("_")[0];
@@ -93,6 +92,10 @@ public class TidyFinancialReportTask {
 						
 						//cp report
 						try {
+							//TODO с╒нд
+							if(reportfile.exists()){
+								continue;
+							}
 							FileUtils.copyFile(reportfile, new File(toDir, filepath));
 						} catch (IOException e) {
 							e.printStackTrace();
